Code quiality improvements according to detekt config.
