+++ app/src/androidTest/java/com/pawel/movieapp/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.pawel.common
+package com.pawel.movieapp
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.common.test", appContext.packageName)
+        assertEquals("com.pawel.movieapp", appContext.packageName)
     }
 }
