--- app/src/androidTest/java/com/pawel/worldline_android_technical_test/CustomTestRunner.kt
@@ -1,18 +1,0 @@
-package com.pawel.worldline_android_technical_test
-
-import android.app.Application
-import android.content.Context
-import androidx.test.runner.AndroidJUnitRunner
-import dagger.hilt.android.testing.HiltTestApplication
-
-@Suppress("unused")
-class CustomTestRunner : AndroidJUnitRunner() {
-
-    override fun newApplication(
-        cl: ClassLoader?,
-        className: String?,
-        context: Context?
-    ): Application {
-        return super.newApplication(cl, HiltTestApplication::class.java.name, context)
-    }
-}
