--- app/src/androidTest/java/com/pawel/worldline_android_technical_test/MatchersHelpers.kt
@@ -1,21 +1,0 @@
-package com.pawel.worldline_android_technical_test
-
-import android.view.View
-import org.hamcrest.Description
-import org.hamcrest.Matcher
-import org.hamcrest.TypeSafeMatcher
-
-fun withIndex(matcher: Matcher<View>, index: Int): Matcher<View> {
-    return object : TypeSafeMatcher<View>() {
-        var currentIndex = 0
-        override fun describeTo(description: Description) {
-            description.appendText("with index: ")
-            description.appendValue(index)
-            matcher.describeTo(description)
-        }
-
-        override fun matchesSafely(view: View?): Boolean {
-            return matcher.matches(view) && currentIndex++ == index
-        }
-    }
-}
