+++ app/src/debug/java/com/pawel/movieapp/HiltTestActivity.kt
@@ -1,0 +1,7 @@
+package com.pawel.movieapp
+
+import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class HiltTestActivity : AppCompatActivity()
