--- app/src/debug/java/com/pawel/worldline_android_technical_test/HiltTestActivity.kt
@@ -1,7 +1,0 @@
-package com.pawel.worldline_android_technical_test
-
-import androidx.appcompat.app.AppCompatActivity
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class HiltTestActivity : AppCompatActivity()
