+++ app/src/main/java/com/pawel/movieapp/FilmAppGlideModule.kt
@@ -1,0 +1,7 @@
+package com.pawel.movieapp
+
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
+
+@GlideModule
+class FilmAppGlideModule : AppGlideModule()
