+++ app/src/main/java/com/pawel/movieapp/MoviesApplication.kt
@@ -1,0 +1,7 @@
+package com.pawel.movieapp
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MoviesApplication : Application()
