--- app/src/main/java/com/pawel/worldline_android_technical_test/FilmAppGlideModule.kt
@@ -1,7 +1,0 @@
-package com.pawel.worldline_android_technical_test
-
-import com.bumptech.glide.annotation.GlideModule
-import com.bumptech.glide.module.AppGlideModule
-
-@GlideModule
-class FilmAppGlideModule : AppGlideModule()
