--- app/src/main/java/com/pawel/worldline_android_technical_test/MoviesApplication.kt
@@ -1,7 +1,0 @@
-package com.pawel.worldline_android_technical_test
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class MoviesApplication : Application()
