+++ common/src/main/java/com/pawel/common/networkErrorHandling/MovieException.kt
@@ -2,7 +2,6 @@ package com.pawel.common.networkErrorHandling
 
 import java.io.IOException
 
-
 const val FUNCTIONNAL_ERROR_CODE = 400
 const val UNAUTHENTICATED_ERROR_CODE = 401
 const val UNAUTHORIZED_ERROR_CODE = 403
