+++ data-mock/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.pawel.worldline_android_technical_test.data_mock">
+    package="com.pawel.movieapp.data_mock">
 
     <application
         tools:targetApi="n"
