+++ data/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.pawel.worldline_android_technical_test.data"/>
+<manifest package="com.pawel.movieapp.data"/>
