+++ data/src/main/java/com/pawel/data/api/MoviesRepositoryNetwork.kt
@@ -2,7 +2,7 @@ package com.pawel.data.api
 
 import com.pawel.domain.model.movie.Movie
 import com.pawel.domain.model.movies.MovieApiResponse
-import com.pawel.worldline_android_technical_test.common.BuildConfig
+import com.pawel.movieapp.common.BuildConfig
 import retrofit2.http.GET
 import retrofit2.http.Path
