+++ presentation/src/main/java/com/pawel/presentation/ui/movieDetail/DetailMovieViewModel.kt
@@ -2,10 +2,7 @@ package com.pawel.presentation.ui.movieDetail
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.pawel.common.networkErrorHandling.MovieException
-import com.pawel.domain.model.movie.Movie
 import com.pawel.domain.service.MoviesService
 import com.pawel.presentation.base.BaseViewModel
 import com.pawel.presentation.helpers.Event
@@ -13,7 +10,6 @@ import com.pawel.presentation.helpers.MoviesError
 import com.pawel.presentation.helpers.NetworkResponse
 import com.pawel.presentation.helpers.SingleMovie
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
